# backend/api/openapi.yaml
openapi: 3.0.0
info:
  title: GitHub Repository Analyzer API
  description: AI-powered code analysis platform for GitHub repositories
  version: 1.0.0
  contact:
    name: API Support
    email: support@acip.com
servers:
  - url: http://localhost:8080
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
...
    CodeChunk:
      type: object
      required:
        - id
        - repositoryId
        - filePath
        - fileName
        - language
        - startLine
        - endLine
        - content
        - createdAt
      properties:
        id:
          type: string
        repositoryId:
          type: string
        filePath:
          type: string
        fileName:
          type: string
        language:
          type: string
        startLine:
          type: integer
          minimum: 1
        endLine:
          type: integer
          minimum: 1
        content:
          type: string
        contentHash:
          type: string
        imports:
          type: array
          items:
            type: string
        metadata:
          type: object
          properties:
            functions:
              type: array
              items:
                type: string
            classes:
              type: array
              items:
                type: string
            variables:
              type: array
              items:
                type: string
            types:
              type: array
              items:
                type: string
            complexity:
              type: integer
              minimum: 0
        vectorId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 500
          description: Search query string
        repositoryId:
          type: string
          description: Optional repository ID filter
        language:
          type: string
          description: Optional programming language filter
        fileType:
          type: string
          description: Optional file extension filter
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
          description: Maximum number of results
        offset:
          type: integer
          minimum: 0
          default: 0
          description: Pagination offset
    SearchResult:
      allOf:
        - $ref: "#/components/schemas/CodeChunk"
        - type: object
          properties:
            score:
              type: number
              format: float
              description: Relevance score
            highlight:
              type: string
              description: Highlighted content snippet
    SearchResponse:
      type: object
      required:
        - results
        - total
        - hasMore
        - query
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        total:
          type: integer
          minimum: 0
          description: Total number of matching results
        hasMore:
          type: boolean
          description: Whether there are more results available
        query:
          type: string
          description: The original search query
    SearchStats:
      type: object
      required:
        - totalChunks
        - totalLines
        - avgComplexity
        - languages
      properties:
        totalChunks:
          type: integer
          minimum: 0
        totalLines:
          type: integer
          minimum: 0
        avgComplexity:
          type: number
          format: float
        languages:
          type: array
          items:
            type: string
...
paths:
...
  /api/search:
    post:
      summary: Search code chunks across all repositories
      description: Perform text search across all accessible code chunks with optional filters
      operationId: globalSearch
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      parameters:
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of results to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: language
          in: query
          description: Filter by programming language
          schema:
            type: string
        - name: fileType
          in: query
          description: Filter by file extension
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/repositories/{id}/search:
    post:
      summary: Search within a specific repository
      description: Perform text search within a specific repository's code chunks
      operationId: repositorySearch
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Repository ID
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of results to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: language
          in: query
          description: Filter by programming language
          schema:
            type: string
        - name: fileType
          in: query
          description: Filter by file extension
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Repository not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/search/quick:
    get:
      summary: Quick search for autocomplete
      description: Lightweight search with limited results for autocomplete functionality
      operationId: quickSearch
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
        - name: repositoryId
          in: query
          description: Filter by repository ID
          schema:
            type: string
        - name: language
          in: query
          description: Filter by programming language
          schema:
            type: string
        - name: fileType
          in: query
          description: Filter by file extension
          schema:
            type: string
      responses:
        "200":
          description: Quick search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        filePath:
                          type: string
                        fileName:
                          type: string
                        language:
                          type: string
                        highlight:
                          type: string
                        score:
                          type: number
                          format: float
                  total:
                    type: integer
                  query:
                    type: string
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/search/suggestions:
    get:
      summary: Get search suggestions
      description: Get autocomplete suggestions for search queries
      operationId: getSearchSuggestions
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Partial search query
          schema:
            type: string
            minLength: 1
        - name: repositoryId
          in: query
          description: Filter by repository ID
          schema:
            type: string
      responses:
        "200":
          description: Search suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestions:
                    type: array
                    items:
                      type: string
                  query:
                    type: string
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/search/languages:
    get:
      summary: Get available programming languages
      description: Retrieve a list of programming languages available for filtering
      operationId: getLanguages
      security:
        - BearerAuth: []
      parameters:
        - name: repositoryId
          in: query
          description: Filter by repository ID
          schema:
            type: string
      responses:
        "200":
          description: Available programming languages
          content:
            application/json:
              schema:
                type: object
                properties:
                  languages:
                    type: array
                    items:
                      type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/search/stats:
    get:
      summary: Get search statistics
      description: Retrieve statistics about the searchable code chunks
      operationId: getSearchStats
      security:
        - BearerAuth: []
      parameters:
        - name: repositoryId
          in: query
          description: Filter by repository ID
          schema:
            type: string
      responses:
        "200":
          description: Search statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/search/recent:
    get:
      summary: Get recently added code chunks
      description: Retrieve the most recently added code chunks
      operationId: getRecentChunks
      security:
        - BearerAuth: []
      parameters:
        - name: repositoryId
          in: query
          description: Filter by repository ID
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of chunks to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Recent code chunks
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunks:
                    type: array
                    items:
                      $ref: "#/components/schemas/CodeChunk"
                  total:
                    type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...